feat: Change PK data type to int

+++ ohouse/src/main/java/com/ohouse/ohouse/entity/Cart.java
@@ -16,7 +16,7 @@ public class Cart {
   @Id
   @GeneratedValue(strategy = GenerationType.IDENTITY)
   @Column(name = "cart_id")
-  private Long cartId;
+  private int cartId;
 
   @ManyToOne(fetch = FetchType.LAZY)
   @JoinColumn(name = "google_id")
